bot/requirements.txt
requests>=2.31.0
httpx>=0.27.0
tweepy>=4.14.0
python-dotenv>=1.0.0
python-telegram-bot>=20.0
stellar-sdk>=10.0.0
